requirements.txt
discord.py[voice]==1.7.3
sounddevice==0.4.2
psutil==5.8.0
pydub==0.25.1
websockets==10.0